// src/generated/intania/auth/student/v1/student.ts
export interface Student {
    /** Internal data */
    id: number;
    createdAt?: Date | undefined;
    updatedAt?: Date | undefined;
    profilePictureKey?: string | undefined;
    /** Student related data */
    studentId?: string | undefined;
    department?: Department | undefined;
    /** Personal data 1 */
    titleTh?: string | undefined;
    titleEn?: string | undefined;
    firstNameTh?: string | undefined;
    firstNameEn?: string | undefined;
    familyNameTh?: string | undefined;
    familyNameEn?: string | undefined;
    middleNameTh?: string | undefined;
    middleNameEn?: string | undefined;
    nicknameTh?: string | undefined;
    nicknameEn?: string | undefined;
    preferredPronoun?: string | undefined;
    nationalId?: string | undefined;
    /** Personal data 2 */
    nationality?: Country | undefined;
    birthDate?: Date | undefined;
    religion?: Religion | undefined;
    bloodType?: string | undefined;
    foodLimitations?: string | undefined;
    drugAllergies?: string | undefined;
    medicalConditions?: string | undefined;
    medications?: string | undefined;
    shirtSize?: number | undefined;
    /** Social */
    email?: string | undefined;
    emailVerified?: boolean | undefined;
    phoneNumber?: string | undefined;
    phoneNumberVerified?: boolean | undefined;
    lineId?: string | undefined;
    facebook?: string | undefined;
    instagram?: string | undefined;
    /** Family */
    familyStatus?: FamilyStatus | undefined;
    /** string enum: "Father", "Mother", "Other" */
    parent?: string | undefined;
    siblingTotal?: number | undefined;
    siblingOrder?: number | undefined;
    parentPhoneNumber?: string | undefined;
    parentAddress?: string | undefined;
    /** Father & Mother */
    fatherName?: string | undefined;
    fatherBirthYear?: number | undefined;
    fatherStatus?: FamilyMemberStatus | undefined;
    motherName?: string | undefined;
    motherBirthYear?: number | undefined;
    motherStatus?: FamilyMemberStatus | undefined;
    /** Current address */
    currentAddressNumber?: string | undefined;
    currentAddressProvince?: Province | undefined;
    currentAddressDistrict?: District | undefined;
    currentAddressLatitude?: number | undefined;
    currentAddressLongitude?: number | undefined;
    currentAddressOther?: string | undefined;
    /** Hometown address */
    hometownAddressNumber?: string | undefined;
    hometownAddressProvince?: Province | undefined;
    hometownAddressDistrict?: District | undefined;
    hometownAddressLatitude?: number | undefined;
    hometownAddressLongitude?: number | undefined;
    hometownAddressOther?: string | undefined;
    /** Miscellaneous */
    cueaDataTransferAgreement?: boolean | undefined;
}

export interface Department {
    id: number;
    createdAt?: Date | undefined;
    updatedAt?: Date | undefined;
    nameTh?: string | undefined;
    nameEn?: string | undefined;
    code?: string | undefined;
}

export interface Country {
    id: number;
    createdAt?: Date | undefined;
    updatedAt?: Date | undefined;
    name?: string | undefined;
    code?: string | undefined;
}

export interface Religion {
    id: number;
    createdAt?: Date | undefined;
    updatedAt?: Date | undefined;
    nameTh?: string | undefined;
    nameEn?: string | undefined;
}

export interface FamilyStatus {
    id: number;
    createdAt?: Date | undefined;
    updatedAt?: Date | undefined;
    valueTh?: string | undefined;
    valueEn?: string | undefined;
}

export interface FamilyMemberStatus {
    id: number;
    createdAt?: Date | undefined;
    updatedAt?: Date | undefined;
    valueTh?: string | undefined;
    valueEn?: string | undefined;
}

export interface Province {
    id: number;
    createdAt?: Date | undefined;
    updatedAt?: Date | undefined;
    nameTh?: string | undefined;
    nameEn?: string | undefined;
    provinceCode?: number | undefined;
}

export interface District {
    id: number;
    createdAt?: Date | undefined;
    updatedAt?: Date | undefined;
    nameTh?: string | undefined;
    nameEn?: string | undefined;
    provinceCode?: number | undefined;
    districtCode?: number | undefined;
    postalCode?: number | undefined;
}
// ...
export const Country: MessageFns<Country> = {
    encode(
        message: Country,
        writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
        if (message.id !== 0) {
            writer.uint32(8).int32(message.id);
        }
        if (message.createdAt !== undefined) {
            Timestamp.encode(
                toTimestamp(message.createdAt),
                writer.uint32(18).fork(),
            ).join();
        }
        if (message.updatedAt !== undefined) {
            Timestamp.encode(
                toTimestamp(message.updatedAt),
                writer.uint32(26).fork(),
            ).join();
        }
        if (message.name !== undefined) {
            writer.uint32(42).string(message.name);
        }
        if (message.code !== undefined) {
            writer.uint32(50).string(message.code);
        }
        return writer;
    },

    decode(input: BinaryReader | Uint8Array, length?: number): Country {
        const reader =
            input instanceof BinaryReader ? input : new BinaryReader(input);
        let end = length === undefined ? reader.len : reader.pos + length;
        const message = createBaseCountry();
        while (reader.pos < end) {
            const tag = reader.uint32();
            switch (tag >>> 3) {
                case 1: {
                    if (tag !== 8) {
                        break;
                    }

                    message.id = reader.int32();
                    continue;
                }
                case 2: {
                    if (tag !== 18) {
                        break;
                    }

                    message.createdAt = fromTimestamp(
                        Timestamp.decode(reader, reader.uint32()),
                    );
                    continue;
                }
                case 3: {
                    if (tag !== 26) {
                        break;
                    }

                    message.updatedAt = fromTimestamp(
                        Timestamp.decode(reader, reader.uint32()),
                    );
                    continue;
                }
                case 5: {
                    if (tag !== 42) {
                        break;
                    }

                    message.name = reader.string();
                    continue;
                }
                case 6: {
                    if (tag !== 50) {
                        break;
                    }

                    message.code = reader.string();
                    continue;
                }
            }
            if ((tag & 7) === 4 || tag === 0) {
                break;
            }
            reader.skip(tag & 7);
        }
        return message;
    },

    fromJSON(object: any): Country {
        return {
            id: isSet(object.id) ? globalThis.Number(object.id) : 0,
            createdAt: isSet(object.createdAt)
                ? fromJsonTimestamp(object.createdAt)
                : undefined,
            updatedAt: isSet(object.updatedAt)
                ? fromJsonTimestamp(object.updatedAt)
                : undefined,
            name: isSet(object.name)
                ? globalThis.String(object.name)
                : undefined,
            code: isSet(object.code)
                ? globalThis.String(object.code)
                : undefined,
        };
    },

    toJSON(message: Country): unknown {
        const obj: any = {};
        if (message.id !== 0) {
            obj.id = Math.round(message.id);
        }
        if (message.createdAt !== undefined) {
            obj.createdAt = message.createdAt.toISOString();
        }
        if (message.updatedAt !== undefined) {
            obj.updatedAt = message.updatedAt.toISOString();
        }
        if (message.name !== undefined) {
            obj.name = message.name;
        }
        if (message.code !== undefined) {
            obj.code = message.code;
        }
        return obj;
    },

    create<I extends Exact<DeepPartial<Country>, I>>(base?: I): Country {
        return Country.fromPartial(base ?? ({} as any));
    },
    fromPartial<I extends Exact<DeepPartial<Country>, I>>(object: I): Country {
        const message = createBaseCountry();
        message.id = object.id ?? 0;
        message.createdAt = object.createdAt ?? undefined;
        message.updatedAt = object.updatedAt ?? undefined;
        message.name = object.name ?? undefined;
        message.code = object.code ?? undefined;
        return message;
    },
};
// ...
export type DeepPartial<T> = T extends Builtin
    ? T
    : T extends globalThis.Array<infer U>
      ? globalThis.Array<DeepPartial<U>>
      : T extends ReadonlyArray<infer U>
        ? ReadonlyArray<DeepPartial<U>>
        : T extends {}
          ? { [K in keyof T]?: DeepPartial<T[K]> }
          : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
    ? P
    : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
          [K in Exclude<keyof I, KeysOfUnion<P>>]: never;
      };
// ...
export interface MessageFns<T> {
    encode(message: T, writer?: BinaryWriter): BinaryWriter;
    decode(input: BinaryReader | Uint8Array, length?: number): T;
    fromJSON(object: any): T;
    toJSON(message: T): unknown;
    create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
    fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}